// frontend/src/feature/types/pokemon.d.ts
type PokemonType =
  | "normal"
  | "fire"
  | "water"
  | "electric"
  | "grass"
  | "ice"
  | "fighting"
  | "poison"
  | "ground"
  | "flying"
  | "psychic"
  | "bug"
  | "rock"
  | "ghost"
  | "dragon"
  | "dark"
  | "steel"
  | "fairy";

type PokemonPackItem = {
  id: string;
  img: string;
  name: string;
};

export { PokemonType, PokemonPackItem };